backend/requirements.txt
asgiref==3.6.0
#backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
DateTimeRange==1.2.0
defusedxml==0.7.1
Django==4.1.5
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==22.1
django-model-utils==4.1.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
importlib-metadata==6.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mbstrdecoder==1.1.1
oauthlib==3.2.2
packaging==23.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typepy==1.3.0
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.11.0
django-shell-plus==1.1.5
# ipython==1.2.1